+++ schema/mutation/deletePage.ts
@@ -15,6 +15,10 @@ export default {
         id: { type: new GraphQLNonNull(GraphQLID) }
     },
     async resolve(parent, args, context) {
+        // Authentication check
+        const user = context.user;
+        if (!user) { throw new GraphQLError(errors.userNotLogged); }
+
         const ability: AppAbility = context.user.ability;
         const filters = Page.accessibleBy(ability, 'delete').where({_id: args.id}).getFilter();
         const page = Page.findOneAndDelete(filters);
