+++ schema/mutation/deleteRole.ts
@@ -13,6 +13,10 @@ export default {
         id: { type: new GraphQLNonNull(GraphQLID) }
     },
     async resolve(parent, args, context) {
+        // Authentication check
+        const user = context.user;
+        if (!user) { throw new GraphQLError(errors.userNotLogged); }
+
         const ability: AppAbility = context.user.ability;
         const filters = Role.accessibleBy(ability, 'delete').where({_id: args.id}).getFilter();
         const role = await Role.findOneAndDelete(filters);
